pyproject.toml
[project]
name = "guardicore-segmentation-response"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil>=7.0.0",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy>=2.0.40",
    "six>=1.17.0",
    "validators>=0.35.0",
    "websocket-client>=1.8.0",
    "python-benedict>=0.34.1",
    "typer>=0.15.3",
    "loguru>=0.7.3",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "faker>=37.1.0",
    "pylint>=3.3.7",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "responses>=0.25.7",
    "ruff>=0.11.8",
]

